requirements.txt
streamlit==1.34.0
pydantic==2.11.4
pdfplumber==0.11.6
python-dotenv==1.1.0
groq==0.4.1
google-generativeai==0.7.1
openai==1.20.0
requests==2.31.0
tqdm==4.67.1
pandas==2.2.2
chromadb==0.4.22
tiktoken==0.5.2
